python: Move error definitions out of a consolidated package to avoid circular imports.
