+++ python/dazl/client/state.py
@@ -6,20 +6,20 @@ from collections import defaultdict
 from typing import Awaitable, Collection, Dict, List, Optional, Union, cast
 import warnings
 
-from ..client._reader_match import is_match
 from ..damlast.daml_lf_1 import TypeConName
 from ..damlast.protocols import SymbolLookup
 from ..model.core import (
     ContractContextualData,
     ContractContextualDataCollection,
-    ContractId,
     ContractMatch,
     ContractsState,
-    UnknownTemplateWarning,
 )
 from ..model.reading import ContractArchiveEvent, ContractCreateEvent
+from ..prim import ContractId
 from ..scheduler import Invoker
 from ..util.asyncio_util import await_then
+from ._reader_match import is_match
+from .errors import UnknownTemplateWarning
 
 
 class ActiveContractSet:
