+++ python/dazl/protocols/autodetect.py
@@ -6,12 +6,12 @@ from threading import Event, RLock, Thread
 from typing import Awaitable, Dict, Iterable, Optional, Union
 
 from .. import LOG
-from ..model.core import ConnectionTimeoutError, UserTerminateRequest
 from ..model.ledger import LedgerMetadata
 from ..model.network import HTTPConnectionSettings
 from ..prim import Party
 from ..scheduler import Invoker
 from ._base import LedgerClient, LedgerConnectionOptions, LedgerNetwork
+from .errors import ConnectionTimeoutError, UserTerminateRequest
 from .oauth import oauth_flow
 from .v1.grpc import GRPCv1Connection
