+++ python/dazl/protocols/v1/grpc.py
@@ -13,7 +13,6 @@ from grpc import Channel, RpcError, insecure_channel, secure_channel, ssl_channe
 
 from ... import LOG
 from ...damlast.daml_lf_1 import PackageRef
-from ...model.core import ConnectionTimeoutError, UserTerminateRequest
 from ...model.ledger import LedgerMetadata
 from ...model.network import HTTPConnectionSettings
 from ...model.reading import BaseEvent, ContractFilter, TransactionFilter
@@ -23,6 +22,7 @@ from ...scheduler import Invoker, RunLevel
 from ...util.io import read_file_bytes
 from ...util.typing import safe_cast
 from .._base import LedgerClient, LedgerConnectionOptions, _LedgerConnection
+from ..errors import ConnectionTimeoutError, UserTerminateRequest
 from .pb_parse_event import (
     BaseEventDeserializationContext,
     serialize_acs_request,
