+++ python/dazl/server/management.py
@@ -7,11 +7,12 @@ Endpoints for managing parties/bots connected via a dazl client.
 
 from dataclasses import asdict, dataclass
 from typing import TYPE_CHECKING, Collection, Sequence
+import warnings
 
 from ..client import Bot, _NetworkImpl
 from ..client.bots import SourceLocation
-from ..model.core import DazlImportError
 from ..prim import Party
+from ..prim.errors import DazlImportError
 
 if TYPE_CHECKING:
     from aiohttp import web
@@ -41,9 +42,11 @@ def build_routes(network_impl: "_NetworkImpl") -> "Collection[web.AbstractRouteD
     try:
         from aiohttp import web
     except ImportError:
-        raise DazlImportError(
-            "aiohttp", "server routes could not be built because aiohttp is not installed"
-        )
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore", DeprecationWarning)
+            raise DazlImportError(
+                "aiohttp", "server routes could not be built because aiohttp is not installed"
+            )
 
     routes = web.RouteTableDef()
